+++ app/src/main/java/com/secretk/move/bean/ProjectHomeBean.java
@@ -471,7 +471,7 @@ public class ProjectHomeBean extends BaseRes{
                 private int evaluationNum;
                 private int discussNum;
                 private int articleNum;
-                private int kffCoinNum;
+                private double kffCoinNum;
                 private Object areaName;
                 private long createTime;
                 private String createTimeStr;
@@ -629,11 +629,11 @@ public class ProjectHomeBean extends BaseRes{
                     this.articleNum = articleNum;
                 }
 
-                public int getKffCoinNum() {
+                public double getKffCoinNum() {
                     return kffCoinNum;
                 }
 
-                public void setKffCoinNum(int kffCoinNum) {
+                public void setKffCoinNum(double kffCoinNum) {
                     this.kffCoinNum = kffCoinNum;
                 }
 
@@ -787,7 +787,7 @@ public class ProjectHomeBean extends BaseRes{
                 private int evaluationNum;
                 private int discussNum;
                 private int articleNum;
-                private int kffCoinNum;
+                private double kffCoinNum;
                 private Object areaName;
                 private long createTime;
                 private String createTimeStr;
@@ -945,11 +945,11 @@ public class ProjectHomeBean extends BaseRes{
                     this.articleNum = articleNum;
                 }
 
-                public int getKffCoinNum() {
+                public double getKffCoinNum() {
                     return kffCoinNum;
                 }
 
-                public void setKffCoinNum(int kffCoinNum) {
+                public void setKffCoinNum(double kffCoinNum) {
                     this.kffCoinNum = kffCoinNum;
                 }
